Port over data path changes for integration, Close futurewei-cloud#36

+++ tools/kernelupdate.sh
@@ -1,27 +1,32 @@
 #!/bin/bash
 
-kernel_dir="../linux-5.6-rc2"
-mkdir -p $kernel_dir
+kernel_ver=`uname -r`
+echo "Running kernel version: $kernel_ver"
 
-wget https://mizar.s3.amazonaws.com/linux-5.6-rc2/linux-headers-5.6.0-rc2_5.6.0-rc2-1_amd64.deb -P $kernel_dir
-wget https://mizar.s3.amazonaws.com/linux-5.6-rc2/linux-image-5.6.0-rc2-dbg_5.6.0-rc2-1_amd64.deb -P $kernel_dir
-wget https://mizar.s3.amazonaws.com/linux-5.6-rc2/linux-image-5.6.0-rc2_5.6.0-rc2-1_amd64.deb -P $kernel_dir
-wget https://mizar.s3.amazonaws.com/linux-5.6-rc2/linux-libc-dev_5.6.0-rc2-1_amd64.deb -P $kernel_dir
+if [ "$kernel_ver" != "5.6.0-rc2" ]; then
+  kernel_dir="linux-5.6-rc2"
+  mkdir -p $kernel_dir
 
-read -p "Continue kernel update (y/n)?" choice
-case "$choice" in
-  y|Y ) echo "Updating kernel";;
-  n|N ) exit;;
-  * ) exit;;
-esac
+  wget https://mizar.s3.amazonaws.com/linux-5.6-rc2/linux-headers-5.6.0-rc2_5.6.0-rc2-1_amd64.deb -P $kernel_dir
+  wget https://mizar.s3.amazonaws.com/linux-5.6-rc2/linux-image-5.6.0-rc2-dbg_5.6.0-rc2-1_amd64.deb -P $kernel_dir
+  wget https://mizar.s3.amazonaws.com/linux-5.6-rc2/linux-image-5.6.0-rc2_5.6.0-rc2-1_amd64.deb -P $kernel_dir
+  wget https://mizar.s3.amazonaws.com/linux-5.6-rc2/linux-libc-dev_5.6.0-rc2-1_amd64.deb -P $kernel_dir
 
-sudo dpkg -i $kernel_dir/*.deb
+  read -p "Continue kernel update (y/n)?" choice
+  case "$choice" in
+    y|Y ) echo "Updating kernel";;
+    n|N ) exit;;
+    * ) exit;;
+  esac
 
-read -p "Reboot host (y/n)?" choice
-case "$choice" in
-  y|Y ) echo "Rebooting";;
-  n|N ) exit;;
-  * ) exit;;
-esac
+  sudo dpkg -i $kernel_dir/*.deb
 
-sudo reboot
+  read -p "Reboot host (y/n)?" choice
+  case "$choice" in
+    y|Y ) echo "Rebooting";;
+    n|N ) exit;;
+    * ) exit;;
+  esac
+
+  sudo reboot
+fi
